+++ mod3rd/__init__.py
@@ -5,7 +5,9 @@ this contains modules not supported by modcore directly
 """
 
 print()
+print("*"*37)
 print( "loading unsupported modules" )
+print("*"*37)
 print()